+++ CHANGELOG.md
@@ -7,6 +7,12 @@ The format is based on [Keep a Changelog]() and this project adheres to
 
 ## [Unreleased]
 
+## [2017-03-02]
+
+### Added
+
+- Working (but unoptimized) Timsort for full sorting.
+
 ## [2017-02-28]
 
 ### Added